@@ -1,5 +1,5 @@
 import React from 'react';
-import LoadingGif from '../../assets/loading.gif';
+import LoadingGif from 'assets/loading.gif';
 import './Loading.css';
 
 export default function Loading() {
