@@ -1,10 +1,10 @@
 import React, { useContext } from 'react';
-import { MoviesShowsContext } from '../../Context';
-import Navbar from '../../components/Navbar/Navbar';
-import ItemCard from '../../components/Item/ItemCard';
-import Loading from '../../components/Loading/Loading';
-import { CONTENT_TYPE } from '../../constants';
-import { IShow, IMovie } from '../../types';
+import { MoviesShowsContext } from 'Context';
+import Navbar from 'components/Navbar/Navbar';
+import ItemCard from 'components/Item/ItemCard';
+import Loading from 'components/Loading/Loading';
+import { CONTENT_TYPE } from 'constantValues';
+import { IShow, IMovie } from 'types';
 import './Home.css';
 
 export default function Home() {
